CAF Aggressors support
Adds support for CAF Aggressors factions as enemy factions. These
factions are not enabled for players on the recommendation of the CAF
team. From their Bohemia forum thread:

"Please keep in mind this mod is created for use as opposing forces only
at this time. The weapons are not designed or tested for player use so
animations and textures may be lacking in quality."